Changes for parser implementation
