--- api2-war/src/main/java/eu/europeana/api2/v2/service/search/syntax/field/FieldMode.java
@@ -1,13 +1,0 @@
-/**
- * 
- */
-package eu.europeana.api2.v2.service.search.syntax.field;
-
-/**
- * @author Hugo
- * @since 7 Feb 2024
- */
-public enum FieldMode
-{
-    search, facet, sort_asc, sort_desc;
-}
