--- api2-war/src/main/java/eu/europeana/api2/v2/service/search/syntax/model/FieldArgumentExpression.java
@@ -1,4 +1,0 @@
-package eu.europeana.api2.v2.service.search.syntax.model;
-
-public interface FieldArgumentExpression extends SyntaxExpression {
-}
