--- api2-war/src/main/java/eu/europeana/api2/v2/service/search/syntax/model/SetExpression.java
@@ -1,4 +1,0 @@
-package eu.europeana.api2.v2.service.search.syntax.model;
-
-public interface SetExpression extends TopLevelExpression, FieldArgumentExpression {
-}
