--- api2-war/src/main/java/eu/europeana/api2/v2/service/search/syntax/model/TopLevelExpression.java
@@ -1,5 +1,0 @@
-package eu.europeana.api2.v2.service.search.syntax.model;
-
-public interface TopLevelExpression extends SyntaxExpression {
-
-}
