+++ api2-war/src/main/java/eu/europeana/api2/v2/web/controller/SearchController.java
@@ -30,9 +30,6 @@ import eu.europeana.api2.v2.model.xml.rss.RssResponse;
 import eu.europeana.api2.v2.service.FacetWrangler;
 import eu.europeana.api2.v2.service.HitMaker;
 import eu.europeana.api2.v2.service.RouteDataService;
-import eu.europeana.api2.v2.service.search.syntax.converter.ConverterContext;
-import eu.europeana.api2.v2.service.search.syntax.model.SyntaxExpression;
-import eu.europeana.api2.v2.service.search.syntax.parser.SearchExpressionParser;
 import eu.europeana.api2.v2.service.translate.TranslationService;
 import eu.europeana.api2.v2.utils.ApiKeyUtils;
 import eu.europeana.api2.v2.utils.BoostParamUtils;
@@ -537,20 +534,22 @@ public class SearchController extends BaseController {
 
     public static void  parseFilterParameter(String newRefinementQuery)
         throws  SolrQueryException {
-        try {
-            if (StringUtils.isNotBlank(newRefinementQuery)) {
-                SearchExpressionParser parser = new SearchExpressionParser(new java.io.StringReader(
-                    newRefinementQuery));
-                SyntaxExpression solrsyntax = parser.parse();
-                LOG.info("### Syntax check passed for filter query ! Parsing completed !####");
-                LOG.info("Solr query:"+   solrsyntax.toSolr(new ConverterContext()));
-
-            }
-        }
-        catch (Exception e){
-            e.printStackTrace();
-            throw new SolrQueryException(ProblemType.SEARCH_QUERY_INVALID);
-        }
+//        try {
+//            if (StringUtils.isNotBlank(newRefinementQuery)) {
+//                SearchExpressionParser parser = new SearchExpressionParser(new java.io.StringReader(
+//                    newRefinementQuery));
+//                SyntaxExpression solrsyntax = parser.parse();
+//                LOG.info("### Syntax check passed for filter query ! Parsing completed !####");
+//
+//                LOG.info("Object model:"+   solrsyntax.toString());
+//                LOG.info("Solr query:"+   solrsyntax.toSolr(new ConverterContext()));
+//
+//            }
+//        }
+//        catch (Exception e){
+//            e.printStackTrace();
+//            throw new SolrQueryException(ProblemType.SEARCH_QUERY_INVALID);
+//        }
     }
 
     /**
