+++ parser/src/main/java/eu/europeana/api/search/syntax/converter/SyntaxConverter.java
@@ -1,0 +1,13 @@
+/**
+ * 
+ */
+package eu.europeana.api.search.syntax.converter;
+
+/**
+ * @author Hugo
+ * @since 7 Feb 2024
+ */
+public interface SyntaxConverter
+{
+    public String toSolr(ConverterContext context);
+}
