+++ parser/src/main/java/eu/europeana/api/search/syntax/field/FieldMode.java
@@ -1,0 +1,13 @@
+/**
+ * 
+ */
+package eu.europeana.api.search.syntax.field;
+
+/**
+ * @author Hugo
+ * @since 7 Feb 2024
+ */
+public enum FieldMode
+{
+    search, facet, sort_asc, sort_desc;
+}
