+++ parser/src/main/java/eu/europeana/api/search/syntax/function/FunctionRegistry.java
@@ -1,0 +1,34 @@
+/**
+ * 
+ */
+package eu.europeana.api.search.syntax.function;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * @author Hugo
+ * @since 7 Feb 2024
+ */
+public class FunctionRegistry {
+
+    public static FunctionRegistry INSTANCE = new FunctionRegistry();
+
+    public Map<String,FunctionClass> registry;
+
+    public FunctionRegistry() {
+        registry = new HashMap();
+    }
+
+    public void addFunction(FunctionClass function) {
+        registry.put(function.getName(), function);
+    }
+
+    public boolean isDeclared(String name) {
+        return registry.containsKey(name);
+    }
+
+    public FunctionClass getFunction(String name) {
+        return registry.get(name);
+    }
+}
