+++ parser/src/main/java/eu/europeana/api/search/syntax/model/ArgumentExpression.java
@@ -1,0 +1,5 @@
+package eu.europeana.api.search.syntax.model;
+
+public interface ArgumentExpression extends SyntaxExpression {
+
+}
