+++ parser/src/main/java/eu/europeana/api/search/syntax/model/ExpressionModelVisitor.java
@@ -1,0 +1,24 @@
+/**
+ * 
+ */
+package eu.europeana.api.search.syntax.model;
+
+/**
+ * @author Hugo
+ * @since 7 Feb 2024
+ */
+public interface ExpressionModelVisitor
+{
+
+    public void visit(AndExpression and);
+
+    public void visit(FieldQueryExpression arg);
+
+    public void visit(FunctionExpression arg);
+
+    public void visit(NotExpression arg);
+
+    public void visit(OrExpression arg);
+
+    public void visit(ValueExpression arg);
+}
