+++ parser/src/main/java/eu/europeana/api/search/syntax/model/FieldArgumentExpression.java
@@ -1,0 +1,4 @@
+package eu.europeana.api.search.syntax.model;
+
+public interface FieldArgumentExpression extends SyntaxExpression {
+}
