+++ parser/src/main/java/eu/europeana/api/search/syntax/model/SetExpression.java
@@ -1,0 +1,4 @@
+package eu.europeana.api.search.syntax.model;
+
+public interface SetExpression extends TopLevelExpression, FieldArgumentExpression {
+}
