+++ parser/src/main/java/eu/europeana/api/search/syntax/model/SyntaxExpression.java
@@ -1,0 +1,15 @@
+/**
+ * 
+ */
+package eu.europeana.api.search.syntax.model;
+
+import eu.europeana.api.search.syntax.converter.SyntaxConverter;
+
+/**
+ * @author Hugo
+ * @since 7 Feb 2024
+ */
+public interface SyntaxExpression extends SyntaxConverter {
+
+    public void visit(ExpressionModelVisitor visitor);
+}
